+++ src/chains/Arbitrum.sol
@@ -5,9 +5,7 @@ import {SettlerBase} from "../SettlerBase.sol";
 import {Settler} from "../Settler.sol";
 import {SettlerMetaTxn} from "../SettlerMetaTxn.sol";
 
-import {IERC20} from "../IERC20.sol";
 import {CurveTricrypto} from "../core/CurveTricrypto.sol";
-import {DodoV1} from "../core/DodoV1.sol";
 import {FreeMemory} from "../utils/FreeMemory.sol";
 
 import {ISettlerActions} from "../ISettlerActions.sol";
@@ -34,7 +32,7 @@ import {SettlerAbstract} from "../SettlerAbstract.sol";
 import {AbstractContext} from "../Context.sol";
 import {Permit2PaymentAbstract} from "../core/Permit2PaymentAbstract.sol";
 
-abstract contract ArbitrumMixin is FreeMemory, SettlerBase, CurveTricrypto, DodoV1 {
+abstract contract ArbitrumMixin is FreeMemory, SettlerBase, CurveTricrypto {
     constructor() {
         assert(block.chainid == 42161 || block.chainid == 31337);
     }
@@ -46,17 +44,7 @@ abstract contract ArbitrumMixin is FreeMemory, SettlerBase, CurveTricrypto, Dodo
         DANGEROUS_freeMemory
         returns (bool)
     {
-        if (super._dispatch(i, action, data)) {
-            return true;
-        } else if (action == ISettlerActions.DODOV1.selector) {
-            (IERC20 sellToken, uint256 bps, address dodo, bool quoteForBase, uint256 minBuyAmount) =
-                abi.decode(data, (IERC20, uint256, address, bool, uint256));
-
-            sellToDodoV1(sellToken, bps, dodo, quoteForBase, minBuyAmount);
-        } else {
-            return false;
-        }
-        return true;
+        return super._dispatch(i, action, data);
     }
 
     function _uniV3ForkInfo(uint8 forkId)