+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Master list of UniV3 forks:
 
 * Added SolidlyV3 UniV3 fork to Base
 * Added SwapBasedV3 UniV3 fork to Base
+* Added support for `DODOV1` action to all chains
+* Added support for `VELODROME` action to all chains
 
 ## 2024-06-27